+++ NEWS.md
@@ -1,3 +1,7 @@
+# cv 2.0.4
+
+- Provide an informative error message when cvCompute() can't refit the model, probably via update() (following report by Joshua Philipp Entrop).
+
 # cv 2.0.3
 
 - New plot.cv() and plot.cvList() methods.